Harden WASM fallback guidance tests
Centralize the GitHub WASM wrapper test setup in src/testing so
wrapper and metadata regressions share one fixture path.

Relax the malformed-first-call assertion to check stable contract
signals instead of the guest's exact validation wording, which keeps
the test resilient if the GitHub schema or parser messages evolve.
