+++ src/tools/wasm/wrapper/metadata.rs
@@ -246,10 +246,9 @@ pub(super) fn build_fallback_guidance(
 mod tests {
     use rstest::{fixture, rstest};
 
-    use crate::testing::{github_wasm_artifact, metadata_test_runtime};
+    use crate::testing::github_wasm_wrapper;
     use crate::tools::Tool;
     use crate::tools::tool::HostedToolCatalogSource;
-    use crate::tools::wasm::capabilities::Capabilities;
 
     use super::super::WasmToolWrapper;
 
@@ -311,15 +310,7 @@ mod tests {
 
     #[fixture]
     async fn github_wrapper() -> WasmToolWrapper {
-        let wasm_path = github_wasm_artifact().expect("build or find github WASM artifact");
-
-        let runtime = metadata_test_runtime().expect("create metadata test runtime");
-        let wasm_bytes = std::fs::read(&wasm_path).expect("read github wasm artifact");
-        let prepared = runtime
-            .prepare("github", &wasm_bytes, None)
-            .await
-            .expect("prepare github wasm component");
-        WasmToolWrapper::new(runtime, prepared, Capabilities::default())
+        github_wasm_wrapper().await
     }
 
     #[rstest]
